# Conflicts:
#	Backend/src/main/java/pe/edu/upc/backend/controllers/PreguntaController.java
#	Backend/src/main/java/pe/edu/upc/backend/serviceinterfaces/IPreguntaService.java

+++ Backend/src/main/java/pe/edu/upc/backend/repositories/IEncuestaPreguntaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pe.edu.upc.backend.entities.EncuestaPregunta;
 @Repository
-public interface IEncuestaPreguntaRepository extends JpaRepository<EncuestaPregunta, Long> {
+public interface IEncuestaPreguntaRepository extends JpaRepository<EncuestaPregunta, Integer> {
 }
