+++ Backend/src/main/java/pe/edu/upc/backend/repositories/IEncuestaRepository.java
@@ -6,6 +6,6 @@ import pe.edu.upc.backend.entities.Encuesta;
 
 
 @Repository
-public interface IEncuestaRepository extends JpaRepository<Encuesta, Long> {
+public interface IEncuestaRepository extends JpaRepository<Encuesta, Integer> {
 
 }
